+++ frontend/package.json
@@ -13,9 +13,12 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.4.0",
     "fullcalendar": "^6.1.7",
+    "react-bootstrap": "^2.7.4",
+    "react-datepicker": "^4.11.0",
     "react-icons": "^4.8.0",
-    "react-router-dom": "^6.11.1",
+    "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
     "styled-components": "^6.0.0-rc.1",
     "web-vitals": "^2.1.4"
